requirements.txt
flask
pillow
pyyaml